Fix line handling and improve monitoring
Reading and writing lines was sometimes unexpectedly causing segfaults
and this was probably due to the way GpioBuffer was allocating native
memory (as less as possible instead of the 64 bytes the native API
allocates).

Also, the index of each line has been encapsulated by creating a new
class (GpioLine). It is not useful for the user to decide the index of a
line and it causes non-obvious problems. For instance, requiring sparse
indexes will result in a failure because the line 0 will be implicitely
requested several times.

When adding an event handle to an event watcher, the user now specifies
an arbitrary identifier. When an event happens, its identifier is given
back to the user. This approach is more flexible than relying on the
number of the line on which the event happened.